No need to instantiate WC_Product_Query multiple times
This commit changes Order::get_random_products() logic to instantiate WC_Product_Query only once instead of instantiating it the number of times equivalent to the number of products that will be returned. As far as I can see it was not necessary to instatiate this class multiple times and to call its get_products() method multiple times, so this change should save some resources while running the command to generate orders.