Reduce number of images created to improve product generation performance
This commit changes the code that generates products to, instead of creating an image gallery for all products, create one only for about 10% of the products. It also changes the highest possible number of images in a gallery from 6 to 3. Creating images is one of the most expensive parts of generating products and this commit reduced the average time spent to create a single product from 2s to 1.5s on my local machine.