Make the methods in the CLI class static for compatibility with PHP 8
Methods in the CLI class are registered using WP_CLI::add_command,
passing array('WC\SmoothGenerator\CLI', 'method') as callbacks.
This kind of callback is not supported in PHP 8 when the method is
not static.

The fix consists of simply making these methods static, which
semantically they are already anyway.