Remove Grunt and related dependencies, run phpcs from composer
The Gruntfile has two commands, one for running the code sniffer (phpcs)
and one for checking the text domain on gettext strings. The latter is
not necessary because the code sniffer can also check text domains.
By running phpcs via a Composer script instead, we can eliminate the
Grunt dependency and all of its sub dependencies entirely, which
reduces the problems associated with npm packages that frequently have
vulnerabilities and need updating.

This also adds the phpcs-changed package and some related scripts
(similar to the WooCommerce Core monorepo) so that the linter will
only check file lines that have been modified instead of giving all
code standard violations for the whole project.