Ensure WC emails are disabled before generating
This consolidates the steps that each generator class needs to take
before generating objects into one `initialize` method. These steps
include creating an instance of the Faker library and unhooking all the
callbacks that send transactional emails. This ensures that the emails
are disabled regardless of which entrypoint to Smooth Generator is used
(cli or web UI).

Fixes #140