Reduce queries while generating orders
This makes a significant reduction in queries by storing lists of
customer and product IDs in a runtime cache, rather than making new
queries for those IDs every time an order is generated. The runtime
cache does get refreshed every 100 orders, so that a greater variety
of customers and products will be used.